+++ spec/lib/xml_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe XmlFu::Markup do
+
+  it "should return a Builder::XmlMarkup object when instantiated" do
+    Builder::XmlMarkup.should === XmlFu::Markup.new
+  end
+end